Move lib/utils/envutils to session/envutils
