+++ lib/srv/reexec.go
@@ -55,11 +55,11 @@ import (
 	"github.com/gravitational/teleport/lib/sshutils/networking"
 	"github.com/gravitational/teleport/lib/sshutils/x11"
 	"github.com/gravitational/teleport/lib/utils"
-	"github.com/gravitational/teleport/lib/utils/envutils"
 	"github.com/gravitational/teleport/lib/utils/host"
 	logutils "github.com/gravitational/teleport/lib/utils/log"
 	"github.com/gravitational/teleport/lib/utils/uds"
 	"github.com/gravitational/teleport/session/auditd"
+	"github.com/gravitational/teleport/session/envutils"
 	"github.com/gravitational/teleport/session/loginuid"
 )
