+++ session/envutils/environment.go
